feat: move package: test -> main (#13)
